+++ src/common/telemetry/index.ts
@@ -1,6 +1,7 @@
 import PnPTelemetry from "@pnp/telemetry-js";
 import { version } from './version';
 import { Environment, EnvironmentType } from "@microsoft/sp-core-library";
+import * as React from 'react';
 
 const CONTROL_TYPE = "react";
 
@@ -20,3 +21,14 @@ export function track(componentName: string, properties: any = {}): void {
     ...properties
   });
 }
+
+export const useTelemetry = (componentName: string, properties: any = {}) => {
+  const [hasBeenCalled, setHasBeenCalled] = React.useState<boolean>(false);
+
+  if (hasBeenCalled) {
+    return;
+  }
+
+  track(componentName, properties);
+  setHasBeenCalled(true);
+};