+++ src/controls/peoplepicker/PeoplePickerComponent.tsx
@@ -93,9 +93,19 @@ export class PeoplePicker extends React.Component<IPeoplePickerProps, IPeoplePic
     if (defaultSelectedUsers) {
       let selectedPersons: IPersonaProps[] = [];
       for (const userValue of props.defaultSelectedUsers) {
-        const userResult = await this.peopleSearchService.searchPersonByEmailOrLogin(userValue, principalTypes, webAbsoluteUrl, this.groupId, ensureUser);
+        let valueAndTitle: string[] = [];
+        valueAndTitle.push(userValue);
+        if (userValue && userValue.indexOf('/') > -1) {
+          valueAndTitle = userValue.split('/');
+        }
+
+        const userResult = await this.peopleSearchService.searchPersonByEmailOrLogin(valueAndTitle[0], principalTypes, webAbsoluteUrl, this.groupId, ensureUser);
         if (userResult) {
           selectedPersons.push(userResult);
+        }
+        else if (valueAndTitle.length === 2 && valueAndTitle[1]) { //user not found.. bind the title if exists
+          const inactiveUser: IPersonaProps = { text: valueAndTitle[1] };
+          selectedPersons.push(inactiveUser);
         }
       }