+++ src/controls/taxonomyPicker/ITaxonomyPicker.ts
@@ -126,6 +126,17 @@ export interface ITaxonomyPickerProps  {
    * Default will be true
    */
   useSessionStorage?: boolean;
+
+  /**
+   * Panel selection change handler. Can be used to interact with the control while selecting items in the panel, before Click or Cancel is clicked.
+   */
+  onPanelSelectionChange?: (prevValue: IPickerTerms, newValue: IPickerTerms) => void;
+
+  /**
+   * Specifies if the childrens should be selected when parent is selected.
+   * By default this is set to false.
+   */
+  selectChildrenIfParentSelected?: boolean;
 }
 
 /**